TLB lookups, indexed writes.

+++ src/memory/constants.rs
@@ -5,11 +5,11 @@
 /// Addresses must be mapped to physical space by the TLB.
 ///
 /// User/supervisor/kernel access.
-pub const USEG_START: usize = 0x0000_0000;
+pub const USEG_START: u32 = 0x0000_0000;
 
-pub const SPRAM_START: usize = 0x7000_0000;
+pub const SPRAM_START: u32 = 0x7000_0000;
 
-pub const USEG_END: usize = KSEG0_START - 1;
+pub const USEG_END: u32 = KSEG0_START - 1;
 
 /// Starting virtual address of kseg0.
 ///
@@ -19,9 +19,9 @@ pub const USEG_END: usize = KSEG0_START - 1;
 /// which contains the BIOS and boot code.
 ///
 /// Kernel access only. Caching controlled by k0 field of Config register (COP0?).
-pub const KSEG0_START: usize = 0x8000_0000;
+pub const KSEG0_START: u32 = 0x8000_0000;
 
-pub const KSEG0_END: usize = KSEG1_START - 1;
+pub const KSEG0_END: u32 = KSEG1_START - 1;
 
 /// Starting virtual address of kseg1.
 ///
@@ -31,9 +31,9 @@ pub const KSEG0_END: usize = KSEG1_START - 1;
 /// which contains the BIOS and boot code.
 ///
 /// Kernel access only. Caching disabled.
-pub const KSEG1_START: usize = 0xA000_0000;
+pub const KSEG1_START: u32 = 0xA000_0000;
 
-pub const KSEG1_END: usize = SSEG_START - 1;
+pub const KSEG1_END: u32 = SSEG_START - 1;
 
 /// Starting virtual address of ksseg/sseg.
 ///
@@ -42,9 +42,9 @@ pub const KSEG1_END: usize = SSEG_START - 1;
 /// Addresses must be mapped to physical space by the TLB.
 ///
 /// Supervisor/kernel access.
-pub const SSEG_START: usize = 0xC000_0000;
+pub const SSEG_START: u32 = 0xC000_0000;
 
-pub const SSEG_END: usize = KSEG3_START - 1;
+pub const SSEG_END: u32 = KSEG3_START - 1;
 
 /// Starting virtual address of kseg3.
 ///
@@ -53,24 +53,24 @@ pub const SSEG_END: usize = KSEG3_START - 1;
 /// Addresses must be mapped to physical space by the TLB.
 ///
 /// Kernel access only.
-pub const KSEG3_START: usize = 0xE000_0000;
+pub const KSEG3_START: u32 = 0xE000_0000;
 
-pub const KSEG3_END: usize = 0xFFFF_FFFF;
+pub const KSEG3_END: u32 = 0xFFFF_FFFF;
 
 // These DMA addresses are courtesy of https://psi-rockin.github.io/ps2tek/.
-pub const IO_REGISTERS_PHYSICAL: usize = 0x1000_0000;
+pub const IO_REGISTERS_PHYSICAL: u32 = 0x1000_0000;
 
-pub const VU0_CODE_PHYSICAL: usize = 0x1100_0000;
-pub const VU0_DATA_PHYSICAL: usize = 0x1100_4000;
-pub const VU1_CODE_PHYSICAL: usize = 0x1100_8000;
-pub const VU1_DATA_PHYSICAL: usize = 0x1100_C000;
+pub const VU0_CODE_PHYSICAL: u32 = 0x1100_0000;
+pub const VU0_DATA_PHYSICAL: u32 = 0x1100_4000;
+pub const VU1_CODE_PHYSICAL: u32 = 0x1100_8000;
+pub const VU1_DATA_PHYSICAL: u32 = 0x1100_C000;
 
-pub const GS_PRIV_REGISTERS_PHYSICAL: usize = 0x1200_0000;
+pub const GS_PRIV_REGISTERS_PHYSICAL: u32 = 0x1200_0000;
 
-pub const IOP_RAM_PHYSICAL: usize = 0x1C00_0000;
+pub const IOP_RAM_PHYSICAL: u32 = 0x1C00_0000;
 
-pub const BIOS_PHYSICAL: usize = 0x1FC0_0000;
-pub const BIOS_START: usize = BIOS_PHYSICAL + KSEG1_START;
+pub const BIOS_PHYSICAL: u32 = 0x1FC0_0000;
+pub const BIOS_START: u32 = BIOS_PHYSICAL + KSEG1_START;
 
 /// Amount of physical memory in the PS2: 32 MiB.
 pub const PHYSICAL_MEMORY_SIZE: usize = 32 * (1 << (10 * 2));
